Aplying Betty

+++ 0x0B-more_malloc_free/0-malloc_checked.c
@@ -1,3 +1,4 @@
+
 #include "holberton.h"
 #include <stdlib.h>
 /**
@@ -8,13 +9,13 @@
  */
 void *malloc_checked(unsigned int b)
 {
-	int *check;
+	long *check;
 
 	check = malloc(b);
 
 	if (check == NULL)
-        {
-	  exit(98);
+	{
+		exit(98);
 	}
 	return (check);
 }
